restructured navigation removed account page the add page now opens with a float button on the main page

+++ app/src/main/java/com/kiologyn/expenda/database/table/record/RecordDao.kt
@@ -28,7 +28,7 @@ interface RecordDao {
         SELECT SUM(record.amount)
         FROM record
     """)
-    suspend fun getBalance(): Int
+    suspend fun getBalance(): Double
 }
 
 data class RecordWithSubcategoryName(
